requirements.txt
streamlit
requests
ai21
dotenv
langchain-ai21
langchain-chroma
langchain-fireworks
langchain-community
